Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    double a = 0.0;
-    double b = 0.0;
+    float a = 0.0;
+    float b = 0.0;
     scanf("%f", &a);
     scanf("%f", &b);
     printf("Product: %.2f", a * b);
